[octopus] removed all size_t types

+++ src/octopus/src/command/CommandQueue.hh
@@ -20,7 +20,7 @@ class CommandData;
 struct CommandBundle
 {
 	CommandVar _var;
-	size_t _idx {0};
+	uint32_t _idx {0};
 };
 
 class CommandQueue
@@ -80,7 +80,7 @@ public:
 	}
 
 	/// @brief Get the bundle associated to the id
-	CommandBundle const *getBundle(size_t id_p) const
+	CommandBundle const *getBundle(uint32_t id_p) const
 	{
 		for(CommandBundle const & bundle_l: _commandQueue)
 		{
@@ -92,7 +92,7 @@ public:
 		return nullptr;
 	}
 	/// @brief Get the bundle associated to the id
-	CommandBundle *getBundle(size_t id_p)
+	CommandBundle *getBundle(uint32_t id_p)
 	{
 		for(CommandBundle & bundle_l: _commandQueue)
 		{
@@ -123,9 +123,9 @@ protected:
 	std::list<CommandBundle> _commandQueue;
 
 	/// @brief used to set up command bundles unique identifiers
-	size_t _idx = 0;
+	uint32_t _idx = 0;
 
-	size_t getNextId()
+	uint32_t getNextId()
 	{
 		return _idx++;
 	}
