+++ src/octopus/src/serialization/CommandSerialization.hh
@@ -30,10 +30,10 @@ void writeDebugCommands(std::ofstream &file_p, Controller const &controller_p);
 /// @brief write a lifst of command to the file
 /// @param file_p
 /// @param list_p
-void writeListOfCommand(std::ofstream &file_p, std::list<Command *> const * list_p, size_t step_l);
+void writeListOfCommand(std::ofstream &file_p, std::list<Command *> const * list_p, uint32_t step_l);
 
 /// @brief Same ad writeListOfCommand but write plain text instead of binary
-void writeDebugListOfCommand(std::ofstream &file_p, std::list<Command *> const * list_p, size_t step_l);
+void writeDebugListOfCommand(std::ofstream &file_p, std::list<Command *> const * list_p, uint32_t step_l);
 
 /// @brief Read commands from a file
 /// @param file_p
